Move Frame GameComponent to a GameEngine Namespace
